Refactor debug JSON helpers. Make _kickRenderLoop use a while loop
