+++ engine/src/flutter/lib/web_ui/lib/src/engine/skwasm/skwasm_impl/renderer.dart
@@ -473,27 +473,6 @@ class SkwasmRenderer extends Renderer {
     );
   }
 
-  String _generateDebugFilename(String filePrefix) {
-    final now = DateTime.now();
-    final String y = now.year.toString().padLeft(4, '0');
-    final String mo = now.month.toString().padLeft(2, '0');
-    final String d = now.day.toString().padLeft(2, '0');
-    final String h = now.hour.toString().padLeft(2, '0');
-    final String mi = now.minute.toString().padLeft(2, '0');
-    final String s = now.second.toString().padLeft(2, '0');
-    return '$filePrefix-$y-$mo-$d-$h-$mi-$s.json';
-  }
-
-  void _dumpDebugInfo(String filePrefix, Map<String, dynamic> json) {
-    final String jsonString = const JsonEncoder.withIndent(' ').convert(json);
-    final blob = createDomBlob([jsonString], {'type': 'application/json'});
-    final url = domWindow.URL.createObjectURL(blob);
-    final element = domDocument.createElement('a');
-    element.setAttribute('href', url);
-    element.setAttribute('download', _generateDebugFilename(filePrefix));
-    element.click();
-  }
-
   @override
   void dumpDebugInfo() {
     if (kDebugMode) {
@@ -530,14 +509,14 @@ class SkwasmRenderer extends Renderer {
           'surfaceCount': counts[26],
           'verticesCount': counts[27],
         };
-        _dumpDebugInfo('live_object_counts', countsJson);
+        downloadDebugInfo('live_object_counts', countsJson);
       });
 
       int i = 0;
       for (final viewRasterizer in rasterizers.values) {
         final Map<String, dynamic>? debugJson = viewRasterizer.dumpDebugInfo();
         if (debugJson != null) {
-          _dumpDebugInfo('flutter-scene$i', debugJson);
+          downloadDebugInfo('flutter-scene$i', debugJson);
           i++;
         }
       }
